requirements.txt
imageio==2.5.0
matplotlib==3.1.1
numpy==1.19.5
opencv-python==4.1.0.25
Pillow==6.1.0
scikit-image==0.15.0
scipy==1.3.1
tensorboard==2.4.0
tensorflow-gpu==2.4.0
tifffile==2020.9.3
Pillow==6.1.0

